fix(pci-instances): hide reverse dns if it is not defined
