+++ packages/manager/apps/pci-instances/src/pages/instances/instance/dashboard/components/InstanceNetworkingBlock.component.tsx
@@ -11,7 +11,6 @@ import {
   ODS_THEME_COLOR_HUE,
   ODS_THEME_COLOR_INTENT,
 } from '@ovhcloud/ods-common-theming';
-import { Text } from '@ovhcloud/ods-react';
 import DashboardCardLayout from './DashboardCardLayout.component';
 import { useProjectId } from '@/hooks/project/useProjectId';
 import { useDashboard } from '../hooks/useDashboard';
@@ -20,28 +19,8 @@ import { BaseActionsMenu } from '@/components/menu/ActionsMenu.component';
 import NetworkItem from './NetworkItem.component';
 import { DashboardTileBlock } from './DashboardTile.component';
 import { useDedicatedUrl } from '@/hooks/url/useDedicatedUrl';
-import { useReverseDns } from '@/data/hooks/network/useNetwork';
 import { useInstanceParams } from '@/pages/instances/action/hooks/useInstanceActionModal';
-
-const ReverseDNS: FC<{ ip: string }> = ({ ip }) => {
-  const { t } = useTranslation('dashboard');
-  const { data, isPending } = useReverseDns(ip);
-
-  return (
-    <LoadingCell isLoading={isPending}>
-      <Text preset="label">{t('pci_instances_dashboard_network_dns')}</Text>
-      {data?.length ? (
-        data.map((dns) => (
-          <div key={dns}>
-            <Text>{dns}</Text>
-          </div>
-        ))
-      ) : (
-        <Text>-</Text>
-      )}
-    </LoadingCell>
-  );
-};
+import ReverseDNS from './ReverseDNS.component';
 
 const InstanceNetworkingBlock: FC = () => {
   const { t } = useTranslation(['dashboard', 'list', 'actions']);
